document_storage_api/requirements.txt
asgiref==3.8.1
boto3==1.35.34
botocore==1.35.34
Django==5.1.1
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
django-cors-headers